+++ example/pure.ts
@@ -1,4 +1,4 @@
 export default {
-    port: 8080,
-    fetch: (request: Request) => new Response('Hi')
+	port: 8080,
+	fetch: (request: Request) => new Response('Hi')
 }